Updates QuizBadgeScrollerViewCell to also use BadgeScrollerItemViewCell
