+++ ThunderCloud/AppCollectionCell.swift
@@ -37,8 +37,6 @@ open class AppCollectionCell: CollectionCell {
         
         super.layoutSubviews()
         collectionView.frame = CGRect(x: 0, y: 1, width: contentView.frame.width, height: 120)
-        pageControl.frame = CGRect(x: 0, y: frame.size.height - 17, width: frame.size.width, height: 12)
-        pageControl.numberOfPages = Int(ceil(collectionView.contentSize.width / collectionView.frame.width))
     }
 }
